fixes, translations

+++ handlers/users/profile.py
@@ -1,4 +1,6 @@
 import logging
+import re
+import traceback
 from aiogram import types
 from aiogram.dispatcher import FSMContext
 from aiogram.dispatcher.handler import CancelHandler
@@ -9,7 +11,6 @@ from data.config import admins
 from database import profile_reg, search_profile
 from database.profile.profile import get_profile_settings_cd, set_profile_settings_cd, update_address
 from filters import IsSubscriber
-from handlers.users.events import return_popup_error
 from keyboards.inline import registration_ikb_menu, yes_no_ikb_menu
 from loader import dp, bot, _
 from states import reger
@@ -24,18 +25,14 @@ from datetime import datetime, timedelta
 async def button_profile_react(message: types.Message, state: FSMContext):
     profile_info = await search_profile("tg_id",message.from_user.id)
     if profile_info != None:
-        await bot.send_message(chat_id=message.from_user.id, text=
-                                 'Ваш профиль:\n'
-                                f'🆔: {profile_info["tg_id"]}\n'
-                                f'👤: {profile_info["game_name"]}\n'
-                                f'💼: {profile_info["role"]}\n'
-                                f'💰: {profile_info["address"]}', reply_markup=edit_profile())
+        profile_out = await get_profile_out(message.from_user.id)
+        await bot.send_message(chat_id=message.from_user.id, text=_('Ваш профиль:\n {profile_out}').format(profile_out=profile_out), reply_markup=edit_profile(),parse_mode="Markdown")
     else:
         a = await bot.send_message(chat_id=message.from_user.id, text=_('👤 Ваш профиль не найден, пройти регистрацию?'),reply_markup=registration_ikb_menu)
         await state.update_data(message_id=a.message_id)
 
 @dp.callback_query_handler(text="Отменить")
-async def disabler(call: CallbackQuery, state: FSMContext):
+async def disabler2(call: CallbackQuery, state: FSMContext):
     data = await state.get_data()
     msgid = data.get("message_id")
     await bot.delete_message(message_id=msgid, chat_id=call.from_user.id)
@@ -46,8 +43,26 @@ async def event_creator(call: CallbackQuery, state: FSMContext):
     msgid = data.get("message_id")
     await bot.edit_message_text(chat_id=call.from_user.id,
                                 message_id=msgid,
-                                text=_('👤 Отправь свою газету с актуальным внутриигровым ником.'))
+                                text=_('👤 Отправь свою газету с актуальным внутриигровым ником.\n(Дата газета не важна, важнен никнейм.)'))
     await reger.tg_id.set()
+
+@dp.callback_query_handler(text="help_bounty", state='*')
+async def disabler(call: CallbackQuery, state: FSMContext):
+    data = await state.get_data()
+    msgid = data.get("message_id")
+    try:
+        a = await bot.send_document(call.from_user.id,
+                            document='https://s2.gifyu.com/images/IMG_2866.gif', parse_mode="HTML")
+        await state.update_data(help_message_id=a.message_id)
+    except:
+        pass
+    
+
+def help():
+    markup = InlineKeyboardMarkup()
+    markup.row_width = 2
+    markup.add(InlineKeyboardButton(_("🆘 Помощь"), callback_data='help_bounty'))
+    return markup
 
 @dp.callback_query_handler(text="bounty_address")
 async def bounty(call: CallbackQuery, state: FSMContext):
@@ -61,55 +76,66 @@ async def bounty(call: CallbackQuery, state: FSMContext):
     if input_datetime >= current_datetime:
         await bot.answer_callback_query(call.id, _('Изменение будет доступно: {input_date}').format(input_date=str(input_datetime), show_alert=True))
         try:
+            profile_out = await get_profile_out(call.from_user.id)
             await bot.edit_message_text(message_id=msgid, chat_id=call.from_user.id,
-                            text=
-                            '❌ Профиль не обновлён.\n'
-                            f'🆔: {profile_info["tg_id"]}\n'
-                            f'👤: {profile_info["game_name"]}\n'
-                            f'💼: {profile_info["role"]}\n'
-                            f'💰: {profile_info["address"]}', reply_markup=edit_profile())
+                            text='⏱ Editing on cooldown.\n'
+                            f'{profile_out}', reply_markup=edit_profile(),parse_mode="Markdown")
         except:
             pass
         CancelHandler()
         await state.finish()
     else:
-        await call.bot.edit_message_text(message_id=call.message.message_id, chat_id=call.from_user.id, text=_('💰 Отправьте вашу геопозицию:\n1. 📎 Прикрепить.\n2. 📍 Геопозиция.\n3. 📍 Отправить геопозицию.\n\n(Следуюшее изменение будет доступно через 31 день)'))
+        await call.bot.edit_message_text(message_id=call.message.message_id, chat_id=call.from_user.id, text=_('💰 Отправьте вашу геопозицию:\n1. 📎 Прикрепить.\n2. 📍 Геопозиция.\n3. 📍 Отправить геопозицию.\n\n(Следуюшее изменение будет доступно через 3 дня)'),reply_markup=help())
         await reger.geopos.set()
-
 
 @dp.message_handler(content_types=ContentTypes.ANY, state=reger.geopos)
 async def state2_reg(message: types.Message, state: FSMContext):
     profile_info = await search_profile("tg_id",message.from_user.id)
+    profile_out = await get_profile_out(message.from_user.id)
     data = await state.get_data()
     msgid = data.get("message_id")
+    help_msg_id = data.get("help_message_id")
     try:
-        url = f'csx://location?lat={message.location.latitude}&lng={message.location.longitude}'
-        await update_address(url, profile_info['tg_id'])
-        cd_date = (datetime.now()+timedelta(days=31))
-        await set_profile_settings_cd('bounty_cd', profile_info['tg_id'], cd_date.strftime("%Y-%m-%d"))
+        await bot.delete_message(chat_id=message.from_user.id, message_id=help_msg_id)
+    except:
+        pass
+    if str(message.content_type) == 'location':
+        try:
+            url = f'csx://location?lat={message.location.latitude}&lng={message.location.longitude}'
+            await update_address(url, profile_info['tg_id'])
+            cd_date = (datetime.now()+timedelta(days=3))
+            await set_profile_settings_cd('bounty_cd', profile_info['tg_id'], cd_date.strftime("%Y-%m-%d"))
+            await bot.delete_message(message_id=message.message_id, chat_id=message.from_user.id)
+            profile_out = await get_profile_out(message.from_user.id)
+            await bot.edit_message_text(message_id=msgid, chat_id=message.from_user.id,
+                                        text=
+                                        '🥳 Profile updated!\n'
+                                        f'{profile_out}', reply_markup=edit_profile(),parse_mode="Markdown")
+        except Exception as e:
+            logging.warning(e)
+            await bot.delete_message(message_id=message.message_id, chat_id=message.from_user.id)
+            await bot.edit_message_text(message_id=msgid, chat_id=message.from_user.id,
+                                text=
+                                '❌ Error while updating.\n'
+                                f'{profile_out}', reply_markup=edit_profile(),parse_mode="Markdown")
+        CancelHandler()
+        await state.finish()
+    else:
         await bot.delete_message(message_id=message.message_id, chat_id=message.from_user.id)
-        profile_info = await search_profile("tg_id",message.from_user.id)
         await bot.edit_message_text(message_id=msgid, chat_id=message.from_user.id,
-                                    text=
-                                    '🥳 Профиль обновлён!\n'
-                                    f'🆔: {profile_info["tg_id"]}\n'
-                                    f'👤: {profile_info["game_name"]}\n'
-                                    f'💼: {profile_info["role"]}\n'
-                                    f'💰: {profile_info["address"]}', reply_markup=edit_profile())
-    except Exception as e:
-        logging.warning(e)
-        await bot.delete_message(message_id=message.message_id, chat_id=message.from_user.id)
-        await bot.edit_message_text(message_id=msgid, chat_id=message.from_user.id,
-                            text=
-                            '❌ Профиль не обновлён.\n'
-                            f'🆔: {profile_info["tg_id"]}\n'
-                            f'👤: {profile_info["game_name"]}\n'
-                            f'💼: {profile_info["role"]}\n'
-                            f'💰: {profile_info["address"]}', reply_markup=edit_profile())
-    CancelHandler()
-    await state.finish()
+                                text=
+                                '❌ Incorrect input.\n'
+                                f'{profile_out}', reply_markup=edit_profile(),parse_mode="Markdown")
+        CancelHandler()
+        await state.finish()
 
-
+async def get_profile_out(tg_id):
+    profile_info = await search_profile("tg_id",tg_id)
+    output = f'🆔: {profile_info["tg_id"]}\n' \
+             f'👤: {profile_info["game_name"]}\n' \
+             f'💼: {profile_info["role"]}\n' \
+             f'💰: `{profile_info["address"]}`'
+    return output
 
 @dp.message_handler(state=reger.tg_id)
 async def state1_reg(message: types.Message, state: FSMContext):
@@ -117,20 +143,29 @@ async def state1_reg(message: types.Message, state: FSMContext):
     data = await state.get_data()
     msgid = data.get("message_id")
     await bot.delete_message(message_id=message.message_id, chat_id=message.from_user.id)
-    if answer == "zxc":
-        answer = "https://press.cheapshot.co/view.html?id=ddb27ad5713275159e522e8a3d909252%2F98ce05f52e8fca5f9ff2e3747d76dc84"
-    if not answer.startswith(('https://press.cheapshot.co', 'http://press.cheapshot.co', 'press.cheapshot.co')):
-        await bot.edit_message_text(message_id=msgid, chat_id=message.from_user.id,text='🚨 Действие отменено\n(Некорректная ссылка)')
+    try:
+        if answer == "zxc":
+            answer = "https://press.cheapshot.co/view.html?id=ddb27ad5713275159e522e8a3d909252%2F98ce05f52e8fca5f9ff2e3747d76dc84"
+        if not answer.startswith(('https://press.cheapshot.co', 'http://press.cheapshot.co', 'press.cheapshot.co')):
+            await bot.edit_message_text(message_id=msgid, chat_id=message.from_user.id,text='🚨 Действие отменено\n(Некорректная ссылка)')
+            CancelHandler()
+            await state.finish()
+        else:
+            answer = re.sub('&inapp=true', '', answer)
+            await bot.edit_message_text(message_id=msgid, chat_id=message.from_user.id,text=f"🕵🏻‍♂️ Читаю газету...")
+            parser = Parser(answer)
+            result = await parser.parse()
+            username = str(result['username'])
+            await bot.edit_message_text(message_id=msgid, chat_id=message.from_user.id,text=f"Вы - `{username}`?",reply_markup=yes_no_ikb_menu, parse_mode='Markdown')
+            await state.update_data(tg_id=message.from_user.id)
+            await state.update_data(name_from_paper=result['username'])
+            await reger.next()
+    except Exception:
+        e = traceback.format_exc()
+        logging.warning(e)
+        await bot.edit_message_text(message_id=msgid, chat_id=message.from_user.id,text='🚨 Действие отменено\n(Ошибка операции)')
         CancelHandler()
         await state.finish()
-    else:
-        await bot.edit_message_text(message_id=msgid, chat_id=message.from_user.id,text=f"🕵🏻‍♂️ Читаю газету...")
-        parser = Parser(answer)
-        result = await parser.parse()
-        await bot.edit_message_text(message_id=msgid, chat_id=message.from_user.id,text=f"Вы - {result['username']}?",reply_markup=yes_no_ikb_menu)
-        await state.update_data(tg_id=message.from_user.id)
-        await state.update_data(name_from_paper=result['username'])
-        await reger.next()
 
 @dp.callback_query_handler(state=reger.name_from_paper)
 async def call1(call: CallbackQuery, state: FSMContext):
@@ -159,7 +194,7 @@ async def call1(call: CallbackQuery, state: FSMContext):
                                     f'🆔: {profile_info["tg_id"]}\n'
                                     f'👤: {profile_info["game_name"]}\n'
                                     f'💼: {profile_info["role"]}\n'
-                                    f'💰: {profile_info["address"]}', reply_markup=edit_profile())
+                                    f'💰: {profile_info["address"]}', reply_markup=edit_profile(), parse_mode='Markdown')
             CancelHandler()
             await state.finish()
         else:
@@ -170,7 +205,6 @@ async def call1(call: CallbackQuery, state: FSMContext):
         CancelHandler()
         await state.finish()
         await call.message.delete()
-
 
 def edit_profile():
     markup = InlineKeyboardMarkup()
